+++ src/standard/webgl/webgl_system.test.ts
@@ -1916,7 +1916,6 @@ describe("WebGLSystem - Render", () => {
                         Polygon.RectangleByDimensions(1, 1, new Vector(2, 0)),
                         new Matrix4D(),
                         new Material({
-                            texture: new Texture("test", Polygon.RectangleByDimensions(1, 1)),
                             shaders: ["test_vert", "test_frag"]
                         }),
                         DrawMode.TRIANGLE_STRIP,
@@ -1952,6 +1951,214 @@ describe("WebGLSystem - Render", () => {
                         new Transform(new Vector(2, 2)),
                         new Camera(undefined, undefined, undefined, new Vector(1, 1))
                     ])]
+                ]),
+                0,
+            ),
+            new Message<number>(Game.MESSAGE_RENDER, 1.0)
+        ],
+        [
+            "Render, unsupported draw mode",
+            undefined,
+            0,
+            new WebGLSystem(
+                new FakeMessageBus(),
+                new FakeWebGL2RenderingContext([
+                    new Reactor("drawArrays", (): void => { drawCount++; }),
+                    new Reactor("canvas", (): HTMLCanvasElement => {
+                        const canvas = new HTMLCanvasElement();
+                        canvas.width = 10;
+                        canvas.height = 10;
+                        return canvas;
+                    }),
+                    new Reactor("createProgram", (): WebGLProgram => new WebGLProgram()),
+                    new Reactor("getProgramParameter", (): number => 1)
+                ]),
+                undefined,
+                [],
+                undefined,
+                new Map<string, [WebGLShader, GLSLShader]>([
+                    ["test_vert", [new WebGLShader(), new GLSLShader(ShaderAsset.VERTEX_TYPE, "")]],
+                    ["test_frag", [new WebGLShader(), new GLSLShader(ShaderAsset.FRAGMENT_TYPE, "")]]
+                ]),
+                new Map<string, WebGLTexture>([
+                    ["test", new WebGLTexture()]
+                ]),
+                new Map<string, WebGLProgram>([
+                    ["_test_vert_test_frag", new WebGLProgram()]
+                ]),
+                new FrustumCuller(new NeverCollideAlgorithm()),
+                new Map<number, SystemEntity>([
+                    [0, new SystemEntity(new FakeEntity(0), [
+                        new Transform(),
+                        new Camera()
+                    ])]
+                ]),
+                0,
+            ),
+            new WebGLSystem(
+                new FakeMessageBus(),
+                new FakeWebGL2RenderingContext([
+                    new Reactor("drawArrays", (): void => { drawCount++; }),
+                    new Reactor("canvas", (): HTMLCanvasElement => {
+                        const canvas = new HTMLCanvasElement();
+                        canvas.width = 10;
+                        canvas.height = 10;
+                        return canvas;
+                    }),
+                    new Reactor("createProgram", (): WebGLProgram => new WebGLProgram()),
+                    new Reactor("getProgramParameter", (): number => 1)
+                ]),
+                undefined,
+                [
+                    new Renderable(0,
+                        Polygon.RectangleByDimensions(1, 1),
+                        new Matrix4D(),
+                        new Material({
+                            texture: new Texture("test", Polygon.RectangleByDimensions(1, 1)),
+                            shaders: ["test_vert", "test_frag"]
+                        }),
+                        23
+                    )
+                ],
+                undefined,
+                new Map<string, [WebGLShader, GLSLShader]>([
+                    ["test_vert", [new WebGLShader(), new GLSLShader(ShaderAsset.VERTEX_TYPE, "")]],
+                    ["test_frag", [new WebGLShader(), new GLSLShader(ShaderAsset.FRAGMENT_TYPE, "")]]
+                ]),
+                new Map<string, WebGLTexture>([
+                    ["test", new WebGLTexture()]
+                ]),
+                new Map<string, WebGLProgram>([
+                    ["_test_vert_test_frag", new WebGLProgram()]
+                ]),
+                new FrustumCuller(new AlwaysCollideAlgorithm()),
+                new Map<number, SystemEntity>([
+                    [0, new SystemEntity(new FakeEntity(0), [
+                        new Transform(),
+                        new Camera()
+                    ])]
+                ]),
+                0,
+            ),
+            new Message<number>(Game.MESSAGE_RENDER, 1.0)
+        ],
+        [
+            "Render, 5 renderables, 1 camera, disjointed z order",
+            undefined,
+            5,
+            new WebGLSystem(
+                new FakeMessageBus(),
+                new FakeWebGL2RenderingContext([
+                    new Reactor("drawArrays", (): void => { drawCount++; }),
+                    new Reactor("canvas", (): HTMLCanvasElement => {
+                        const canvas = new HTMLCanvasElement();
+                        canvas.width = 10;
+                        canvas.height = 10;
+                        return canvas;
+                    }),
+                    new Reactor("createProgram", (): WebGLProgram => new WebGLProgram()),
+                    new Reactor("getProgramParameter", (): number => 1)
+                ]),
+                undefined,
+                [],
+                undefined,
+                new Map<string, [WebGLShader, GLSLShader]>([
+                    ["test_vert", [new WebGLShader(), new GLSLShader(ShaderAsset.VERTEX_TYPE, "", () => { return; }, () => { return; }, () => { return; })]],
+                    ["test_frag", [new WebGLShader(), new GLSLShader(ShaderAsset.FRAGMENT_TYPE, "", () => { return; }, () => { return; }, () => { return; })]]
+                ]),
+                new Map<string, WebGLTexture>([
+                    ["test", new WebGLTexture()]
+                ]),
+                new Map<string, WebGLProgram>([
+                    ["_test_vert_test_frag", new WebGLProgram()]
+                ]),
+                new FrustumCuller(new AlwaysCollideAlgorithm()),
+                new Map<number, SystemEntity>([
+                    [0, new SystemEntity(new FakeEntity(0), [
+                        new Transform(new Vector(1, 1)),
+                        new Camera(undefined, undefined, undefined, new Vector(1, 1))
+                    ])],
+                ]),
+                0,
+            ),
+            new WebGLSystem(
+                new FakeMessageBus(),
+                new FakeWebGL2RenderingContext([
+                    new Reactor("drawArrays", (): void => { drawCount++; }),
+                    new Reactor("canvas", (): HTMLCanvasElement => {
+                        const canvas = new HTMLCanvasElement();
+                        canvas.width = 10;
+                        canvas.height = 10;
+                        return canvas;
+                    }),
+                    new Reactor("createProgram", (): WebGLProgram => new WebGLProgram()),
+                    new Reactor("getProgramParameter", (): number => 1)
+                ]),
+                undefined,
+                [
+                    new Renderable(0,
+                        Polygon.RectangleByDimensions(1, 1, new Vector(0, 0)),
+                        new Matrix4D(),
+                        new Material({
+                            texture: new Texture("test", Polygon.RectangleByDimensions(1, 1)),
+                            shaders: ["test_vert", "test_frag"]
+                        }),
+                        DrawMode.TRIANGLE_STRIP
+                    ),
+                    new Renderable(1,
+                        Polygon.RectangleByDimensions(1, 1, new Vector(1, 0)),
+                        new Matrix4D(),
+                        new Material({
+                            texture: new Texture("test", Polygon.RectangleByDimensions(1, 1)),
+                            shaders: ["test_vert", "test_frag"]
+                        }),
+                        DrawMode.TRIANGLE_STRIP
+                    ),
+                    new Renderable(0,
+                        Polygon.RectangleByDimensions(1, 1, new Vector(2, 0)),
+                        new Matrix4D(),
+                        new Material({
+                            texture: new Texture("test", Polygon.RectangleByDimensions(1, 1)),
+                            shaders: ["test_vert", "test_frag"]
+                        }),
+                        DrawMode.TRIANGLE_STRIP
+                    ),
+                    new Renderable(3,
+                        Polygon.RectangleByDimensions(1, 1, new Vector(3, 0)),
+                        new Matrix4D(),
+                        new Material({
+                            texture: new Texture("test", Polygon.RectangleByDimensions(1, 1)),
+                            shaders: ["test_vert", "test_frag"]
+                        }),
+                        DrawMode.TRIANGLE_STRIP
+                    ),
+                    new Renderable(5,
+                        Polygon.RectangleByDimensions(1, 1, new Vector(4, 0)),
+                        new Matrix4D(),
+                        new Material({
+                            texture: new Texture("test", Polygon.RectangleByDimensions(1, 1)),
+                            shaders: ["test_vert", "test_frag"]
+                        }),
+                        DrawMode.TRIANGLE_STRIP
+                    )
+                ],
+                undefined,
+                new Map<string, [WebGLShader, GLSLShader]>([
+                    ["test_vert", [new WebGLShader(), new GLSLShader(ShaderAsset.VERTEX_TYPE, "", () => { return; }, () => { return; }, () => { return; })]],
+                    ["test_frag", [new WebGLShader(), new GLSLShader(ShaderAsset.FRAGMENT_TYPE, "", () => { return; }, () => { return; }, () => { return; })]]
+                ]),
+                new Map<string, WebGLTexture>([
+                    ["test", new WebGLTexture()]
+                ]),
+                new Map<string, WebGLProgram>([
+                    ["_test_vert_test_frag", new WebGLProgram()]
+                ]),
+                new FrustumCuller(new AlwaysCollideAlgorithm()),
+                new Map<number, SystemEntity>([
+                    [0, new SystemEntity(new FakeEntity(0), [
+                        new Transform(new Vector(1, 1)),
+                        new Camera(undefined, undefined, undefined, new Vector(1, 1))
+                    ])],
                 ]),
                 0,
             ),