+++ CHANGELOG.md
@@ -18,7 +18,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
     - Trilinear.
   - Can specify generating mipmaps.
 - Primitive rendering, can now render primitive shapes by adding a `Primitive`
-  component - can render lines, polygons, and points; allows specifying draw mode.
+  component - can render lines, polygons, and points; allows specifying draw
+  mode.
+  - Allows more low level control over rendering, rather than `Sprite`, which
+    has a fixed draw mode and is further processed by the `SpriteSystem`.
+- `Material` property `color` added, if there is a texture the texture will be
+  mixed with this color, if there is no texture the color will be used directly.
 ### Changed
 - Images now loaded with a new `ImageRequest` rather than an `ImageAsset` -
   `ImageAsset` used internally.
@@ -26,7 +31,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Optional font options specified by a `IFontOptions` interface, allows
   overriding default font options such as `characters`, `buffer`, `cutoff` etc.
 - `Material` options specified by new `IMaterialOptions`, a collection of
-  optional arguments that override material defaults.
+  optional arguments that override material defaults, such as `texture`, `color`
+  and `shaders`.
+- `Material` property `texture` now optional, if none provided it will render
+  only using the `color`.
 ### Fixed
 - Z order bug where an error was thrown if there were gaps between z orders,
   e.g. `0, 1, 3` would error because `2` was missing - caused due to internal